% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gamma_relative}
\alias{gamma_relative}
\title{Return relative transition probabilities}
\usage{
gamma_relative(gamma)
}
\arguments{
\item{gamma}{transition probability matrix}
}
\value{
Relative transition probability matrix
}
\description{
Return relative transition probabilities
}
