% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{mHMM_AIC}
\alias{mHMM_AIC}
\title{Compute AIC for a \code{mHMM_list} object}
\usage{
mHMM_AIC(models)
}
\arguments{
\item{models}{mHMM_list opject containing multiple chains of a multilevel hidden markov model.}
}
\value{
Named vector containing loglikelihood, AIC, and AICc
}
\description{
Compute AIC for a \code{mHMM_list} object
}
