% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stat_dist}
\alias{stat_dist}
\title{Obtain stationary distribution of a transition probability matrix}
\usage{
stat_dist(gamma)
}
\arguments{
\item{gamma}{}
}
\value{
Vector with marginal probabilities.
}
\description{
Obtain stationary distribution of a transition probability matrix
}
