% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vinkphd-package.R
\docType{package}
\name{vinkphd-package}
\alias{vinkphd}
\alias{vinkphd-package}
\title{vinkphd: Utility Package for PhD}
\description{
Package with utility functions that I use in my PhD.
}
\author{
\strong{Maintainer}: Pepijn Vink \email{p.a.vink@uu.nl}

}
\keyword{internal}
