% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vit_mHMM_list.R
\name{vit_mHMM_list}
\alias{vit_mHMM_list}
\title{Obtain hidden state sequence for each subject using the Viterbi
algorithm}
\usage{
vit_mHMM_list(models, s_data, burn_in = NULL, return_state_prob = FALSE)
}
\arguments{
\item{models}{An object of class \code{mHMM}, generated
by the function \code{\link{mHMM}}.}

\item{s_data}{Data file.}

\item{burn_in}{The number of iterations to be discarded from the MCMC
algorithm when inferring the transition probability matrix gamma and the
emission distribution of (each of) the dependent variable(s) for each
subject from \code{s_data}. If omitted, defaults to \code{NULL} and
\code{burn_in} specified at the function \code{mHMM} will be used.}

\item{return_state_prob}{A logical scaler. Should the function, in addition
to the most likely state sequence for each subject, also return the state
probabilities at each point in time for each subject
(\code{return_state_prob = TRUE}) or not (\code{return_state_prob =
  FALSE}). Defaults to \code{return_state_prob = FALSE}.}
}
\value{
The function \code{vit_mHMM} returns a matrix containing the most
likely state at each point in time. The first column indicates subject id
number. Hence, the id number is repeated over rows equal to the number of
observations for that subject. The second column contains the most likely
state at each point in time. If requested, the subsequent columns contain
the state probabilities at each point in time for each subject.
}
\description{
\code{vit_mHMM} obtains the most likely state sequence (for each subject)
from an object of class \code{mHMM} (generated by the function
\code{mHMM()}), using (an extended version of) the Viterbi algorithm. This is
also known as global decoding. Optionally, the state probabilities
themselves at each point in time can also be obtained.
}
\details{
Note that local decoding is also possible, by inferring the most frequent
state at each point in time for each subject from the sampled state path at
each iteration of the MCMC algorithm. This information is contained in the
output object \code{return_path} of the function \code{mHMM()}.
}
